+++ gapic-common/lib/gapic/call_options/retry_policy.rb
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 module Gapic
-  class ApiCall
+  class CallOptions
     ##
     # The policy for retrying failed API calls using an incremental backoff. A new object instance should be used for
     # every ApiCall invocation.
