+++ gapic-common/test/gapic/grpc/api_call/api_call_test.rb
@@ -23,13 +23,13 @@ class ApiCallTest < Minitest::Test
       OperationStub.new { 42 }
     end
 
-    api_call = Gapic::ApiCall.new api_meth_stub
+    api_call = Gapic::Grpc::Stub::ApiCall.new api_meth_stub
 
     assert_equal 42, api_call.call(Object.new)
     assert_kind_of Time, deadline_arg
 
     new_deadline = Time.now + 20
-    options = Gapic::ApiCall::Options.new timeout: 20
+    options = Gapic::CallOptions.new timeout: 20
 
     assert_equal 42, api_call.call(Object.new, options: options)
     assert_in_delta new_deadline, deadline_arg, 0.9
@@ -42,7 +42,7 @@ class ApiCallTest < Minitest::Test
     end
 
     format_response = ->(response) { response.to_s }
-    api_call = Gapic::ApiCall.new api_meth_stub
+    api_call = Gapic::Grpc::Stub::ApiCall.new api_meth_stub
 
     assert_equal 5, api_call.call(3)
     assert_equal "5", api_call.call(3, format_response: format_response)
@@ -58,7 +58,7 @@ class ApiCallTest < Minitest::Test
     end
 
     increment_addr = ->(*args) { adder = 5 }
-    api_call = Gapic::ApiCall.new api_meth_stub
+    api_call = Gapic::Grpc::Stub::ApiCall.new api_meth_stub
 
     assert_equal 5, api_call.call(3)
     assert_equal 5, api_call.call(3, operation_callback: increment_addr)
@@ -75,7 +75,7 @@ class ApiCallTest < Minitest::Test
 
     format_response = ->(response) { response.to_s }
     increment_addr = ->(*args) { adder = 5 }
-    api_call = Gapic::ApiCall.new api_meth_stub
+    api_call = Gapic::Grpc::Stub::ApiCall.new api_meth_stub
 
     assert_equal 5, api_call.call(3)
     assert_equal "5", api_call.call(3, format_response: format_response, operation_callback: increment_addr)
@@ -93,7 +93,7 @@ class ApiCallTest < Minitest::Test
     end
 
     collect_response = ->(response) { all_responses << response }
-    api_call = Gapic::ApiCall.new api_meth_stub
+    api_call = Gapic::Grpc::Stub::ApiCall.new api_meth_stub
 
     api_call.call([:foo, :bar, :baz].to_enum, stream_callback: collect_response)
     wait_until { all_responses == [:foo, :bar, :baz] }
@@ -113,7 +113,7 @@ class ApiCallTest < Minitest::Test
 
     collect_response = ->(response) { all_responses << response }
     format_response = ->(response) { response.to_s }
-    api_call = Gapic::ApiCall.new api_meth_stub
+    api_call = Gapic::Grpc::Stub::ApiCall.new api_meth_stub
 
     api_call.call([:foo, :bar, :baz].to_enum, stream_callback: collect_response)
     wait_until { all_responses == [:foo, :bar, :baz] }
@@ -131,7 +131,7 @@ class ApiCallTest < Minitest::Test
       OperationStub.new { requests.each(&block) }
     end
 
-    api_call = Gapic::ApiCall.new api_meth_stub
+    api_call = Gapic::Grpc::Stub::ApiCall.new api_meth_stub
 
     responses = api_call.call [:foo, :bar, :baz].to_enum
     assert_kind_of Enumerable, responses
@@ -153,7 +153,7 @@ class ApiCallTest < Minitest::Test
     end
 
     format_responses = ->(responses) { responses.lazy.map(&:to_s) }
-    api_call = Gapic::ApiCall.new api_meth_stub
+    api_call = Gapic::Grpc::Stub::ApiCall.new api_meth_stub
 
     responses = api_call.call [:foo, :bar, :baz].to_enum
     assert_kind_of Enumerable, responses
