+++ gapic-common/test/gapic/paged_enumerable/invalid_response_test.rb
@@ -18,7 +18,7 @@ class PagedEnumerableInvalidResponseTest < Minitest::Test
   def test_MissingRepeatedResponse
     request = Gapic::Examples::GoodPagedRequest.new
     response = Gapic::Examples::MissingRepeatedResponse.new
-    options = Gapic::ApiCall::Options.new
+    options = Gapic::CallOptions.new
 
     error = assert_raises ArgumentError do
       Gapic::PagedEnumerable.new(
@@ -32,7 +32,7 @@ class PagedEnumerableInvalidResponseTest < Minitest::Test
   def test_MissingMessageResponse
     request = Gapic::Examples::GoodPagedRequest.new
     response = Gapic::Examples::MissingMessageResponse.new
-    options = Gapic::ApiCall::Options.new
+    options = Gapic::CallOptions.new
 
     error = assert_raises ArgumentError do
       Gapic::PagedEnumerable.new(
@@ -46,7 +46,7 @@ class PagedEnumerableInvalidResponseTest < Minitest::Test
   def test_MissingNextPageTokenResponse
     request = Gapic::Examples::GoodPagedRequest.new
     response = Gapic::Examples::MissingNextPageTokenResponse.new
-    options = Gapic::ApiCall::Options.new
+    options = Gapic::CallOptions.new
 
     error = assert_raises ArgumentError do
       Gapic::PagedEnumerable.new(
@@ -62,7 +62,7 @@ class PagedEnumerableInvalidResponseTest < Minitest::Test
 
     request = Gapic::Examples::GoodPagedRequest.new
     response = Gapic::Examples::BadMessageOrderResponse.new
-    options = Gapic::ApiCall::Options.new
+    options = Gapic::CallOptions.new
 
     error = assert_raises ArgumentError do
       Gapic::PagedEnumerable.new(
