+++ gapic-generator/templates/default/service/client/method/docs/_request_streaming.erb
@@ -1,5 +1,5 @@
 <%- assert_locals method -%>
 # @param request [Gapic::StreamInput, Enumerable<<%= method.request_type %> | Hash>]
 #   An enumerable of {<%= method.request_type %>} instances.
-# @param options [Gapic::ApiCall::Options, Hash]
+# @param options [Gapic::CallOptions, Hash]
 #   Overrides the default settings for this call, e.g, timeout, retries, etc. Optional.
